+++ PorteriaV3/src/java/Controllers/MenuPorteriaController.java
@@ -109,7 +109,7 @@ public class MenuPorteriaController implements Serializable {
         }
     }
 
-    public MenuPorteria getMenuPorteria(java.lang.Long id) {
+    public MenuPorteria getMenuPorteria(java.lang.Integer id) {
         return getFacade().find(id);
     }
 
@@ -134,13 +134,13 @@ public class MenuPorteriaController implements Serializable {
             return controller.getMenuPorteria(getKey(value));
         }
 
-        java.lang.Long getKey(String value) {
-            java.lang.Long key;
-            key = Long.valueOf(value);
+        java.lang.Integer getKey(String value) {
+            java.lang.Integer key;
+            key = Integer.valueOf(value);
             return key;
         }
 
-        String getStringKey(java.lang.Long value) {
+        String getStringKey(java.lang.Integer value) {
             StringBuilder sb = new StringBuilder();
             sb.append(value);
             return sb.toString();
